+++ src/db/Database.ts
@@ -274,6 +274,7 @@ export const spaceIndicator = Settings.settings.space.value;
 export const showLines = Settings.settings.lines.value;
 export const showAnnotations = Settings.settings.annotations.value;
 export const mic = Settings.settings.mic.value;
+export const voice = Settings.settings.say.value;
 export const blocks = Settings.settings.blocks.value;
 export const howToNotifications = Settings.settings.howToNotifications.value;
 export const status = DB.Status;